added insert customer by button and update datatable and grow

+++ src/main/java/org/acme/springjsf/repository/CustomerRepository.java
@@ -1,10 +1,8 @@
 package org.acme.springjsf.repository;
 
-import java.math.BigInteger;
-
 import org.acme.springjsf.model.Customer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CustomerRepository extends JpaRepository<Customer, BigInteger> {
+public interface CustomerRepository extends JpaRepository<Customer, Integer> {
 
 }
